Feature/updated c code

+++ QuickTCGA/Logic/NucleusSeg_Yi/ProcessTileUtils.h
@@ -12,9 +12,9 @@
 #include <opencv2/opencv.hpp>
 
 
-cv::Mat processTile(cv::Mat thisTileCV, float otsuRatio = 1.0, double curvatureWeight = 0.8, float sizeThld = 3, float sizeUpperThld = 200, double mpp = 0.25, double ks = 15.0);
+cv::Mat processTile(cv::Mat thisTileCV, float otsuRatio = 1.0, double curvatureWeight = 0.8, float sizeThld = 3, float sizeUpperThld = 200, double mpp = 0.25, float ks = 15.0);
 
-cv::Mat processTileNoDeclump(cv::Mat thisTileCV, float otsuRatio = 1.0, double curvatureWeight = 0.8, float sizeThld = 3, float sizeUpperThld = 200, double mpp = 0.25, double ks = 15.0);
+cv::Mat processTileNoDeclump(cv::Mat thisTileCV, float otsuRatio = 1.0, double curvatureWeight = 0.8, float sizeThld = 3, float sizeUpperThld = 200, double mpp = 0.25, float ks = 15.0);
 
 //template< typename itkImage_t > void writeImage(typename itkImage_t::Pointer img, const char *fileName, bool compress);
